+++ src/test/java/org/cicirello/ds/FibonacciMinHeapTests.java
@@ -35,6 +35,21 @@ public class FibonacciMinHeapTests extends SharedTestHelpersMinHeaps {
 	}
 	
 	@Test
+	public void testDefaultMinHeap() {
+		defaultMinHeap();
+	}
+	
+	@Test
+	public void testDefaultReverseMinHeap() {
+		defaultReverseMinHeap();
+	}
+	
+	@Test
+	public void testDefaultArbitraryMinHeap() {
+		defaultArbitraryMinHeap();
+	}
+	
+	@Test
 	public void testElementPollThenAddMinHeap() {
 		elementPollThenAddMinHeap();
 	}
