+++ src/test/java/org/cicirello/ds/SimpleFibonacciMinHeapTests.java
@@ -35,6 +35,21 @@ public class SimpleFibonacciMinHeapTests extends SharedTestHelpersMinHeaps {
 	}
 	
 	@Test
+	public void testDefaultMinHeap() {
+		defaultDuplicatesAllowedMinHeap();
+	}
+	
+	@Test
+	public void testDefaultReverseMinHeap() {
+		defaultReverseDuplicatesAllowedMinHeap();
+	}
+	
+	@Test
+	public void testDefaultArbitraryMinHeap() {
+		defaultArbitraryDuplicatesAllowedMinHeap();
+	}
+	
+	@Test
 	public void testElementPollThenAddMinHeap() {
 		elementPollThenAddMinHeap();
 	}
